Rename masked_input to masked_number_input
