+++ assets/js/hooks/index.js
@@ -4,4 +4,4 @@ export { default as BackpexSidebarSections } from './_sidebar_sections'
 export { default as BackpexStickyActions } from './_sticky_actions'
 export { default as BackpexThemeSelector } from './_theme_selector'
 export { default as BackpexTooltip } from './_tooltip'
-export { default as BackpexMaskedInput } from './_masked_input'
+export { default as BackpexMaskedNumberInput } from './_masked_number_input'
